FederatedFogSimulationV4/data_collector/requirements.txt
paho-mqtt
matplotlib